techmo_sarmata_pyclient/requirements.txt
grpcio==1.1.3
grpcio-tools==1.1.3
protobuf==3.1.0
pydub
